# midas-backend/public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"claims": {
						"items": {
							"$ref": "#/components/schemas/Claim"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"claims",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Claim": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"contact_number": {
						"type": "string"
					},
					"user_id": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"client_id": {
						"type": "number",
						"format": "double"
					},
					"client": {
						"$ref": "#/components/schemas/Client"
					},
					"receivers": {
						"items": {
							"$ref": "#/components/schemas/Receiver"
						},
						"type": "array"
					},
					"tasks": {
						"items": {
							"$ref": "#/components/schemas/Task"
						},
						"type": "array"
					},
					"documents": {
						"items": {
							"$ref": "#/components/schemas/Document"
						},
						"type": "array"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"contact_number",
					"user_id",
					"user",
					"client_id",
					"client",
					"receivers",
					"tasks",
					"documents",
					"start_date",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Client": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"access_code": {
						"type": "string"
					},
					"claim_id": {
						"type": "number",
						"format": "double"
					},
					"claim": {
						"$ref": "#/components/schemas/Claim"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"access_code",
					"claim_id",
					"claim",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Receiver": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"access_code": {
						"type": "string"
					},
					"claim_id": {
						"type": "number",
						"format": "double"
					},
					"claim": {
						"$ref": "#/components/schemas/Claim"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"access_code",
					"claim_id",
					"claim",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Task": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"claim_id": {
						"type": "number",
						"format": "double"
					},
					"claim": {
						"$ref": "#/components/schemas/Claim"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"completed_date": {
						"type": "string",
						"format": "date-time"
					},
					"deadline_date": {
						"type": "string",
						"format": "date-time"
					},
					"details": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"claim_id",
					"claim",
					"start_date",
					"completed_date",
					"deadline_date",
					"details",
					"type",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRecordOfAny": {
				"properties": {},
				"type": "object",
				"additionalProperties": {}
			},
			"Document": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"uploader_name": {
						"type": "string"
					},
					"upload_url": {
						"type": "string"
					},
					"claim_id": {
						"type": "number",
						"format": "double"
					},
					"claim": {
						"$ref": "#/components/schemas/Claim"
					},
					"type": {
						"type": "string"
					},
					"job_id": {
						"type": "string"
					},
					"form_data": {
						"items": {
							"$ref": "#/components/schemas/IRecordOfAny"
						},
						"type": "array"
					},
					"raw": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"title",
					"status",
					"uploader_name",
					"upload_url",
					"claim_id",
					"claim",
					"type",
					"job_id",
					"form_data",
					"raw",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DeleteResult": {
				"description": "Result object returned by DeleteQueryBuilder execution.",
				"properties": {
					"raw": {
						"description": "Raw SQL result returned by executed query."
					},
					"affected": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "Number of affected rows/documents\nNot all drivers support this"
					}
				},
				"required": [
					"raw"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IClientPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"claim_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"email",
					"claim_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IClientVerifyPayload": {
				"properties": {
					"email": {
						"type": "string"
					},
					"access_code": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"access_code"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IDocumentPayload": {
				"properties": {
					"uploader_name": {
						"type": "string"
					},
					"claim_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"uploader_name",
					"claim_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IS3Payload": {
				"properties": {
					"parent": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"data": {
						"items": {
							"properties": {
								"job_id": {
									"type": "string"
								},
								"name": {
									"type": "string"
								}
							},
							"required": [
								"job_id",
								"name"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"parent",
					"type",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IReceiverPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"claim_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"email",
					"claim_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IReceiverVerifyPayload": {
				"properties": {
					"email": {
						"type": "string"
					},
					"access_code": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"access_code"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "midas-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/clients": {
			"get": {
				"operationId": "GetClients",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Client"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateClient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Client"
								}
							}
						}
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IClientPayload"
							}
						}
					}
				}
			}
		},
		"/clients/{id}": {
			"get": {
				"operationId": "GetClient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Client"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "UpdateClient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Client"
								}
							}
						}
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IClientPayload"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteClient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/DeleteResult"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/clients/verify": {
			"post": {
				"operationId": "VerifyClient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Client"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Client"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IClientVerifyPayload"
							}
						}
					}
				}
			}
		},
...
		"/documents/{id}": {
			"post": {
				"operationId": "UpdateDocument",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Document"
								}
							}
						}
					}
				},
				"tags": [
					"Document"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IDocumentPayload"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetDocument",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Document"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Document"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteDocument",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/DeleteResult"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Document"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/documents/claim/{id}": {
			"get": {
				"operationId": "GetClaimDocuments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Document"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Document"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/documents/download/{id}": {
			"get": {
				"operationId": "DownloadDocument",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Document"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/documents/presignedUrl/{id}": {
			"get": {
				"operationId": "GenerateDocumentPresignedUrl",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Document"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/documents/bootstrap": {
			"post": {
				"operationId": "CreateS3Document",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Document"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IS3Payload"
							}
						}
					}
				}
			}
		},
		"/receivers": {
			"get": {
				"operationId": "GetReceivers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Receiver"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Receiver"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateReceiver",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Receiver"
								}
							}
						}
					}
				},
				"tags": [
					"Receiver"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IReceiverPayload"
							}
						}
					}
				}
			}
		},
		"/receivers/{id}": {
			"get": {
				"operationId": "GetReceiver",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Receiver"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Receiver"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "UpdateReceiver",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Receiver"
								}
							}
						}
					}
				},
				"tags": [
					"Receiver"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IReceiverPayload"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteReceiver",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/DeleteResult"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Receiver"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/receivers/claim/{id}": {
			"get": {
				"operationId": "GetReceiversByClaim",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Receiver"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Receiver"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/receivers/verify": {
			"post": {
				"operationId": "VerifyReceiver",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Receiver"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Receiver"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IReceiverVerifyPayload"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}